# docs/swagger.yaml
basePath: /
definitions:
  common.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
  model.Container:
    properties:
      address:
        type: string
      cmd:
        items:
          type: string
        type: array
      id:
        type: string
      image:
        type: string
      name:
        type: string
      port:
        type: integer
      startAt:
        type: string
      status:
        type: string
    type: object
  model.CreatedContainer:
    properties:
      cmd:
        items:
          type: string
        type: array
      image:
        type: string
      name:
        type: string
      port:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This is a weave server api doc.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Weave Server API
  version: "2.0"
paths:
...
  /api/v1/containers:
    get:
      description: List container
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.Container'
                  type: array
              type: object
      security:
      - JWT: []
      summary: List container
      tags:
      - container
    post:
      consumes:
      - application/json
      description: Create container
      parameters:
      - description: container info
        in: body
        name: container
        required: true
        schema:
          $ref: '#/definitions/model.CreatedContainer'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Container'
              type: object
      security:
      - JWT: []
      summary: Create container
      tags:
      - container
  /api/v1/containers/:id:
    put:
      consumes:
      - application/json
      description: Update container
      parameters:
      - description: container info
        in: body
        name: container
        required: true
        schema:
          $ref: '#/definitions/model.CreatedContainer'
      - description: container id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Container'
              type: object
      security:
      - JWT: []
      summary: Update container
      tags:
      - container
  /api/v1/containers/{id}:
    delete:
      description: Delete container
      parameters:
      - description: container id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - JWT: []
      summary: Delete container
      tags:
      - container
    get:
      description: Get container
      parameters:
      - description: container id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Container'
              type: object
      security:
      - JWT: []
      summary: Get container
      tags:
      - container
    post:
      description: Operate container
      parameters:
      - description: container id
        in: path
        name: id
        required: true
        type: string
      - description: 'verb: start/stop/restart'
        in: query
        name: verb
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - JWT: []
      summary: Operate container
      tags:
      - container
  /api/v1/containers/{id}/exec:
    get:
      description: exec container
      parameters:
      - description: container id
        in: path
        name: id
        required: true
        type: string
      - description: shell, sh or bash
        in: query
        name: shell
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - JWT: []
      summary: Exec container
      tags:
      - container
  /api/v1/containers/{id}/log:
    get:
      description: Get container log
      parameters:
      - description: container id
        in: path
        name: id
        required: true
        type: string
      - description: follow log
        in: query
        name: follow
        type: boolean
      - description: tail log all or number
        in: query
        name: tail
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - JWT: []
      summary: Get container log
      tags:
      - container
  /api/v1/containers/{id}/proxy:
    get:
      description: proxy container
      parameters:
      - description: container id
        in: path
        name: id
        required: true
        type: string
      - description: shell, sh or bash
        in: query
        name: shell
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - JWT: []
      summary: Proxy container
      tags:
      - container
...
securityDefinitions:
  JWT:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"